refactor:清理patchouli无效测试
