+++ tests/integration/patchouli/test_memory_generation_pipeline.py
