+++ tests/unit/patchouli/control/test_memory_generation_queue.py
@@ -144,18 +144,6 @@ def test_spec_codec_creates_canonical_deep_snapshot_and_restores_domain_types() 
     assert second.spec.request.context.state_summary == "original summary"
 
 
-def test_active_identity_uses_stable_intent_task_id() -> None:
-    work = _MemoryGenerationWork(
-        task_id="active:intent-codec",
-        spec=_spec(intent_id="intent-codec", pending_alias="draft-codec"),
-    )
-
-    identity = _MemoryGenerationWorkAdapter.identity(work)
-
-    assert identity.work_id == "memory_generation:active:intent-codec"
-    assert identity.idempotency_key == "active:intent-codec"
-
-
 @pytest.mark.asyncio
 async def test_concurrency_limit_keeps_later_task_queued_and_pending() -> None:
     first_started = asyncio.Event()
