+++ tests/unit/patchouli/memory_library/test_buffer.py
@@ -43,18 +43,6 @@ class TestSemanticBufferCreation:
         buf = SemanticBuffer()
         assert buf.total_tokens == 0
 
-    def test_custom_fields(self):
-        buf = SemanticBuffer(
-            topic_id="custom_topic",
-            user_id="user_123",
-            topic_title="Custom Title",
-            current_agent_id="agent_a",
-        )
-        assert buf.topic_id == "custom_topic"
-        assert buf.user_id == "user_123"
-        assert buf.topic_title == "Custom Title"
-        assert buf.current_agent_id == "agent_a"
-
 
 class TestSemanticBufferClear:
     """SemanticBuffer.clear() 方法测试"""
@@ -82,10 +70,14 @@ class TestSemanticBufferClear:
     def test_clear_updates_last_update_timestamp(self):
         buf = SemanticBuffer()
         old_timestamp = buf.last_update
+        fixed_now = datetime(2027, 1, 1, 12, 0, 0)
 
-        buf.clear()
+        with patch("hivememory.patchouli.memory_library.buffer.datetime") as mock_datetime:
+            mock_datetime.now.return_value = fixed_now
+            buf.clear()
 
-        assert buf.last_update >= old_timestamp
+        assert buf.last_update == fixed_now.timestamp()
+        assert buf.last_update > old_timestamp
 
 
 class TestSemanticBufferBlockCount:
