+++ tests/unit/patchouli/services/test_lifecycle.py
@@ -2,12 +2,8 @@
 LifecycleFamiliar 单元测试
 
 测试覆盖:
-- run_gardening_once: 全局维护调度器入口
-- refresh_memory_vitality: 批量生命力刷新
-- record_hit: 命中事件记录
-- record_citation: 引用事件记录
-- record_feedback: 用户反馈记录
-- revive_memory: 长期记忆复活
+- run_gardening_once: 全局维护调度器入口（编排合约 + 异常处理）
+- record_hit: 字符串 ID 归一化为 UUID
 """
 
 import pytest
@@ -51,72 +47,6 @@ class TestLifecycleFamiliar:
         assert result["success"] is False
         assert "GC failed" in result["error"]
         assert result["archived_count"] == 0
-
-    @pytest.mark.asyncio
-    async def test_refresh_memory_vitality_delegates_to_engine(self):
-        from hivememory.core.models import MemoryAtom
-        lifecycle = Mock()
-        lifecycle.refresh_vitality_batch = AsyncMock(return_value=[(uuid4(), 1.0)])
-        familiar = self._make_familiar(lifecycle=lifecycle)
-        memory = Mock(spec=MemoryAtom)
-
-        result = await familiar.refresh_memory_vitality([memory], persist=True)
-
-        lifecycle.refresh_vitality_batch.assert_awaited_once_with([memory], persist=True)
-
-    @pytest.mark.asyncio
-    async def test_record_hit_delegates_to_engine(self):
-        lifecycle = Mock()
-        lifecycle.record_hit = AsyncMock(return_value={"memory_id": "id", "delta": 0.1})
-        familiar = self._make_familiar(lifecycle=lifecycle)
-        memory_id = uuid4()
-
-        result = await familiar.record_hit(memory_id, source="retrieval")
-
-        lifecycle.record_hit.assert_awaited_once()
-        call_args = lifecycle.record_hit.call_args
-        assert call_args[0][0] == memory_id
-        assert call_args[1]["source"] == "retrieval"
-
-    @pytest.mark.asyncio
-    async def test_record_citation_delegates_to_engine(self):
-        lifecycle = Mock()
-        lifecycle.record_citation = AsyncMock(return_value={"memory_id": "id", "delta": 0.05})
-        familiar = self._make_familiar(lifecycle=lifecycle)
-        memory_id = uuid4()
-
-        result = await familiar.record_citation(memory_id, source="mtp")
-
-        lifecycle.record_citation.assert_awaited_once()
-        call_args = lifecycle.record_citation.call_args
-        assert call_args[0][0] == memory_id
-        assert call_args[1]["source"] == "mtp"
-
-    @pytest.mark.asyncio
-    async def test_record_feedback_delegates_to_engine(self):
-        lifecycle = Mock()
-        lifecycle.record_feedback = AsyncMock(return_value={"memory_id": "id", "delta": 0.2})
-        familiar = self._make_familiar(lifecycle=lifecycle)
-        memory_id = uuid4()
-
-        result = await familiar.record_feedback(memory_id, positive=True, source="ui")
-
-        lifecycle.record_feedback.assert_awaited_once()
-        call_args = lifecycle.record_feedback.call_args
-        assert call_args[0][0] == memory_id
-        assert call_args[1]["positive"] is True
-        assert call_args[1]["source"] == "ui"
-
-    @pytest.mark.asyncio
-    async def test_revive_memory_calls_library_revive(self):
-        memory_id = uuid4()
-        memory_lib = Mock()
-        memory_lib.revive = AsyncMock()
-        familiar = self._make_familiar(memory_lib=memory_lib)
-
-        await familiar.revive_memory(memory_id)
-
-        memory_lib.revive.assert_awaited_once_with(memory_id)
 
     @pytest.mark.asyncio
     async def test_normalize_uuid_converts_string_to_uuid(self):
