+++ tests/unit/patchouli/services/test_perception.py
@@ -2,14 +2,16 @@
 PerceptionFamiliar 单元测试
 
 测试覆盖:
-- submit_interaction: 交互摄入与 settlement 提交
+- submit_interaction: 交互摄入与 settlement 提交（mock 边界）
 - manual_settle_topic: 手动结算
 - evict_topic: 话题驱逐
 - discard_if_empty: 空话题清理
 - scan_idle_buffers_once: 空闲话题扫描与结算
+
+真实链路（PerceptionFamiliar + Layer + Store + TriggerManager 协作）测试
+位于 tests/integration/patchouli/test_perception_flush_chain.py。
 """
 
-import time
 from types import SimpleNamespace
 from unittest.mock import AsyncMock, Mock
 
@@ -18,78 +20,11 @@ import pytest
 from hivememory.core.models import Identity, LogicalBlock, TopicData, TurnEvent, TurnRecord
 from hivememory.core.protocol.models import InteractionPayload
 from hivememory.engines.perception.models import FlushReason, TopicMaterializeTask
-from hivememory.engines.perception.semantic_flow_perception_layer import (
-    SemanticFlowPerceptionLayer,
-)
 from hivememory.patchouli.contracts.local_routes import PatchouliLocalRoutes
 from hivememory.patchouli.control.interaction_apply_journal import (
     InMemoryInteractionApplyJournal,
 )
-from hivememory.patchouli.memory_library.library import MemoryLibrary
-from hivememory.patchouli.memory_library.stores import ShortTermMemoryStore
 from hivememory.patchouli.services.perception import PerceptionFamiliar
-from hivememory.system.config import SemanticFlowPerceptionConfig
-
-
-def _make_identity(user="u1", agent="a1"):
-    return Identity(user_id=user, agent_id=agent)
-
-
-def _make_payload(user_msg="hello", assistant_msg="world", identity=None):
-    identity = identity or _make_identity()
-    return InteractionPayload(
-        user_message=user_msg,
-        assistant_final_text=assistant_msg,
-        turn_events=[
-            TurnEvent(
-                kind="assistant_message",
-                sequence=0,
-                role="assistant",
-                content=assistant_msg,
-            )
-        ],
-        identity=identity,
-    )
-
-
-def _make_real_familiar(
-    *,
-    idle_timeout_seconds=1,
-    max_resident_topics=5,
-    fold_token_threshold=999999,
-    fold_retain_recent_blocks=2,
-):
-    store = ShortTermMemoryStore(max_resident_topics=max_resident_topics)
-    relay = Mock()
-    relay.should_relay.return_value = None
-    relay.generate_summary.side_effect = (
-        lambda blocks_to_fold, previous_summary: (
-            f"{previous_summary}|folded:{len(blocks_to_fold)}"
-            if previous_summary
-            else f"folded:{len(blocks_to_fold)}"
-        )
-    )
-    interaction_journal = InMemoryInteractionApplyJournal()
-    layer = SemanticFlowPerceptionLayer(
-        config=SemanticFlowPerceptionConfig(
-            fold_token_threshold=fold_token_threshold,
-            fold_retain_recent_blocks=fold_retain_recent_blocks,
-        ),
-        relay_controller=relay,
-        short_term_store=store,
-        interaction_journal=interaction_journal,
-    )
-    bus = Mock()
-    bus.request = AsyncMock(return_value=None)
-    library = MemoryLibrary(short_term=store, mid_term=Mock(), long_term=Mock())
-    familiar = PerceptionFamiliar(
-        perception_layer=layer,
-        bus=bus,
-        config=SimpleNamespace(idle_timeout_seconds=idle_timeout_seconds),
-        memory_library=library,
-        interaction_journal=interaction_journal,
-    )
-    return familiar, layer, store, bus
 
 
 class TestPerceptionFamiliar:
@@ -312,8 +247,12 @@ class TestPerceptionFamiliar:
 
         result = await familiar.manual_settle_topic()
 
-        assert result is expected_task
-        bus.request.assert_awaited_once()
+        assert result is not None
+        layer.settle_topic.assert_awaited_once_with("t1", FlushReason.MANUAL)
+        assert (
+            bus.request.await_args.args[0]
+            == PatchouliLocalRoutes.GENERATION_SUBMIT_SETTLEMENT
+        )
 
     @pytest.mark.asyncio
     async def test_evict_topic_calls_layer_swap_out(self):
@@ -326,17 +265,6 @@ class TestPerceptionFamiliar:
 
         assert result.success is True
         layer.swap_out_topic.assert_called_once_with("topic_to_evict")
-
-    def test_discard_if_empty_delegates_to_layer(self):
-        """验证 discard_if_empty 正确委托给 layer"""
-        layer = Mock()
-        layer.discard_if_empty = Mock(return_value=True)
-        familiar = self._make_familiar(layer=layer)
-
-        result = familiar.discard_if_empty("empty_topic")
-
-        assert result is True
-        layer.discard_if_empty.assert_called_once_with("empty_topic")
 
     @pytest.mark.asyncio
     async def test_scan_idle_buffers_once_skips_non_idle_topics(self):
@@ -356,114 +284,9 @@ class TestPerceptionFamiliar:
         store.list_topic_data = Mock(return_value=[recent_topic_data])
 
         bus = Mock()
-        familiar = self._make_familiar(layer=layer, store=store, bus=bus)
-        familiar._idle_timeout_seconds = 3600  # 1小时
+        familiar = self._make_familiar(layer=layer, store=store, bus=bus, idle_timeout=3600)
 
         flushed = await familiar.scan_idle_buffers_once()
 
         assert len(flushed) == 0
         layer.settle_topic.assert_not_called()
-
-    @pytest.mark.asyncio
-    async def test_idle_flush_swaps_out_topic(self):
-        familiar, _, store, bus = _make_real_familiar(idle_timeout_seconds=1)
-        await familiar.submit_interaction(_make_payload("question", "answer"), "NEW_TOPIC")
-        assert len(store.list_topic_data()) == 1
-
-        time.sleep(1.1)
-        flushed = await familiar.scan_idle_buffers_once()
-
-        assert len(flushed) == 1
-        assert store.list_topic_data() == []
-        bus.request.assert_awaited_with(
-            PatchouliLocalRoutes.GENERATION_SUBMIT_SETTLEMENT,
-            bus.request.await_args.args[1],
-        )
-        assert bus.request.await_args.args[1].reason == FlushReason.IDLE_TIMEOUT
-
-    @pytest.mark.asyncio
-    async def test_idle_flush_skips_empty_settlement_submission(self):
-        familiar, layer, store, bus = _make_real_familiar(idle_timeout_seconds=1)
-        topic_id = await layer.create_new_topic(_make_identity())
-        assert store.get_topic_data(topic_id) is not None
-
-        time.sleep(1.1)
-        flushed = await familiar.scan_idle_buffers_once()
-
-        assert flushed == [topic_id]
-        bus.request.assert_not_awaited()
-
-    @pytest.mark.asyncio
-    async def test_idle_flush_frees_slot(self):
-        familiar, _, store, _ = _make_real_familiar(
-            idle_timeout_seconds=1,
-            max_resident_topics=2,
-        )
-        await familiar.submit_interaction(_make_payload("q1", "a1", _make_identity("u1", "a1")), "NEW_TOPIC")
-        await familiar.submit_interaction(_make_payload("q2", "a2", _make_identity("u2", "a2")), "NEW_TOPIC")
-        assert len(store.list_topic_data()) == 2
-
-        time.sleep(1.1)
-        assert len(await familiar.scan_idle_buffers_once()) == 2
-
-        await familiar.submit_interaction(_make_payload("q3", "a3", _make_identity("u3", "a3")), "NEW_TOPIC")
-        assert len(store.list_topic_data()) == 1
-
-    @pytest.mark.asyncio
-    async def test_shutdown_flush_archives_and_swaps_out_all_topics(self):
-        familiar, _, store, bus = _make_real_familiar(max_resident_topics=4)
-        await familiar.submit_interaction(_make_payload("q1", "a1", _make_identity("u1", "a1")), "NEW_TOPIC")
-        await familiar.submit_interaction(_make_payload("q2", "a2", _make_identity("u2", "a2")), "NEW_TOPIC")
-
-        result = await familiar.flush_all_for_shutdown()
-
-        assert result.trigger_reason == FlushReason.SHUTDOWN.value
-        assert len(result.flushed_topics) == 2
-        assert result.archived_blocks == 2
-        assert store.list_topic_data() == []
-        assert bus.request.await_count == 2
-        for call in bus.request.await_args_list:
-            assert call.args[0] == PatchouliLocalRoutes.GENERATION_SUBMIT_SETTLEMENT
-            assert call.args[1].reason == FlushReason.SHUTDOWN
-
-    @pytest.mark.asyncio
-    async def test_shutdown_after_folding_settles_summary_and_retained_block(self):
-        familiar, _, store, bus = _make_real_familiar(
-            max_resident_topics=2,
-            fold_token_threshold=10,
-            fold_retain_recent_blocks=1,
-        )
-        identity = _make_identity("u-fold", "a-fold")
-        topic_id = "NEW_TOPIC"
-
-        for i in range(3):
-            topic_id = await familiar.submit_interaction(
-                _make_payload(
-                    f"question-{i}-" * 80,
-                    f"answer-{i}",
-                    identity,
-                ),
-                topic_id,
-            )
-
-        before_shutdown = store.get_topic_data(topic_id, touch=False)
-        assert before_shutdown is not None
-        assert before_shutdown.state_summary == "folded:1|folded:1"
-        assert [block.user_query for block in before_shutdown.blocks] == [
-            "question-2-" * 80
-        ]
-        assert bus.request.await_count == 0
-
-        result = await familiar.flush_all_for_shutdown()
-
-        assert result.flushed_topics == [topic_id]
-        assert result.archived_blocks == 1
-        assert store.get_topic_data(topic_id, touch=False) is None
-        bus.request.assert_awaited_once()
-        route, settlement = bus.request.await_args.args
-        assert route == PatchouliLocalRoutes.GENERATION_SUBMIT_SETTLEMENT
-        assert settlement.reason == FlushReason.SHUTDOWN
-        assert settlement.state_summary == "folded:1|folded:1"
-        assert [block.user_query for block in settlement.blocks] == [
-            "question-2-" * 80
-        ]
